refactor: separate hooks
